Feature/ccip 632 rate limiter gas optimization (#803)
* optimize rate limiter gas by using more packed structs

* fix ABI

* gen wrappers

* update offchain

* remove 64bit rate limiter

* reoreder rate limit config structs

* address comments

* address comments

* extra new tokens calc into own function

* properly reproduce bug in test

* address comments

* rename RateLimiter128 back to RateLimiter

* add comments

* better comments for rate limiter

* better comments for rate limiter

+++ contracts/src/v0.8/ccip/AggregateRateLimiter.sol
@@ -31,7 +31,7 @@ contract AggregateRateLimiter is OwnerIsCreator {
       rate: config.rate,
       capacity: config.capacity,
       tokens: config.capacity,
-      lastUpdated: uint40(block.timestamp),
+      lastUpdated: uint32(block.timestamp),
       isEnabled: config.isEnabled
     });
   }
